# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-commerce API Documentation",
        "description": "API documentation for the e-commerce platform",
        "contact": {
            "email": "karimashraf851@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "List all users",
                "operationId": "d7a57f4d7de18fb0f92501172472c9b6",
                "responses": {
                    "200": {
                        "description": "List of users"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "operationId": "9fea6f2adcb69614c521a7733358382d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{user}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user details",
                "operationId": "8ef94a31920758a2ea8c4a11c4a85c73",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user details",
                "operationId": "d8774c2ad802c372cefa655ff2fe989d",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete a user",
                "operationId": "d2769c9ccb9342ab54329ddb86aa438c",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "NotFoundError": {
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "StoreUserRequest": {
                "required": [
                    "name",
                    "email",
                    "password",
                    "password_confirmation",
                    "role"
                ],
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/NotFoundError"
                    },
                    {
                        "properties": {
                            "name": {
                                "type": "string",
                                "maxLength": 255
                            },
                            "email": {
                                "type": "string",
                                "format": "email",
                                "maxLength": 255
                            },
                            "password": {
                                "type": "string",
                                "format": "password",
                                "minLength": 8
                            },
                            "password_confirmation": {
                                "type": "string",
                                "format": "password"
                            },
                            "role": {
                                "type": "string",
                                "enum": [
                                    "admin",
                                    "customer"
                                ]
                            },
                            "image": {
                                "type": "string",
                                "format": "binary"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "UpdateUserRequest": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/NotFoundError"
                    },
                    {
                        "properties": {
                            "name": {
                                "type": "string",
                                "maxLength": 255
                            },
                            "email": {
                                "type": "string",
                                "format": "email",
                                "maxLength": 255
                            },
                            "password": {
                                "type": "string",
                                "format": "password",
                                "minLength": 8
                            },
                            "role": {
                                "type": "string",
                                "enum": [
                                    "admin",
                                    "customer"
                                ]
                            },
                            "image": {
                                "type": "string",
                                "format": "binary"
                            },
                            "is_admin": {
                                "type": "boolean"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}